fix(desktop): update playground shape — RangeSlider/Label primitives, controls extracted under 400 lines
